requirements.txt
Babel==2.3.4
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
decorator==4.0.10
docutils==0.12
ebaysdk==2.1.5
feedparser==5.2.1
geographiclib==1.49
geopy==1.18.1
gevent==1.1.2
greenlet==0.4.10
html2text==2016.9.19
idna==2.7
Jinja2==2.8.1
libsass==0.12.3
lxml==3.7.1
Mako==1.0.4
MarkupSafe==0.23
mock==2.0.0
num2words==0.5.6
ofxparse==0.16
olefile==0.46
passlib==1.6.5
pbr==5.1.1
Pillow==5.4.1
psutil==4.3.1
psycopg2==2.7.3.1
py3dns==3.2.0
pydot==1.2.3
pyfcm==1.4.5
PyJWT==1.7.1
pyldap==2.4.28
pyotp==2.2.7
pyparsing==2.1.10
PyPDF2==1.26.0
pyserial==3.1.1
PySocks==1.6.8
python-dateutil==2.5.3
python-stdnum==1.10
pytz==2016.7
pyusb==1.0.0
qrcode==5.3
reportlab==3.3.0
requests==2.20.0
requests-toolbelt==0.9.1
six==1.12.0
soupsieve==1.7.2
suds-jurko==0.6
twilio==6.24.0
urllib3==1.24.1
validate-email==1.3
vatnumber==1.2
vobject==0.9.3
Werkzeug==0.11.15
xlrd==1.0.0
XlsxWriter==0.9.3
xlwt==1.3.0
